My portifolio project in Python
In this Python script, I created a few functions to analyze the data in the "Insurance.csv" dataset based on specific goals I had in mind for this project. These were simple goals, and I might come back to this project in the future for refinement. I've also attached my Jupyter notebook for this project.